# frontend/swagger.yaml
openapi: 3.0.3
info:
  title: YoudaoNoteLM API
  description: |
    有道云笔记 LM 后端服务 API 文档。

    ## 认证方式
    除认证模块（`/auth/*`）外，所有接口均需在 Header 中携带 JWT Access Token：
    ```
    Authorization: Bearer <access_token>
    ```

    ## 统一响应格式
    ```json
    {
      "code": 0,
      "message": "成功",
      "data": {}
    }
    ```
    - `code = 0` 表示成功
    - `code != 0` 为业务错误码，详见错误码表
  version: 1.0.0
  contact:
    name: YoudaoNoteLM Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token
  schemas:
...
    Response:
      type: object
      properties:
        code:
          type: integer
          description: 业务错误码，0=成功
          example: 0
        message:
          type: string
          example: 成功
        data:
          description: 响应数据
...
    URLImportRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://example.com/article
    SearchImportRequest:
      type: object
      required: [urls]
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          example:
            - https://example.com/article1
            - https://example.com/article2
    TaskIDResponse:
      type: object
      properties:
        task_id:
          type: string
          example: uuid-string

    # ========== 错误码 ==========
...
paths:
...
  /notebooks/{nbId}/search/url:
    post:
      tags: [Search]
      summary: URL 直接导入
      description: 抓取指定 URL 的内容并导入到笔记本
      operationId: importFromURL
      security:
        - BearerAuth: []
      parameters:
        - name: nbId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLImportRequest'
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskIDResponse'
  /notebooks/{nbId}/search/import:
    post:
      tags: [Search]
      summary: 批量导入搜索结果
      description: 批量抓取多个 URL 并导入，返回任务 ID 用于查询进度
      operationId: importSearchResults
      security:
        - BearerAuth: []
      parameters:
        - name: nbId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchImportRequest'
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskIDResponse'

  # ==================== Health ====================
...